+++ lib/toast_pack.dart
@@ -1,7 +1,7 @@
 /// A lightweight, zero-dependency toast package for Flutter.
 library;
 
-export 'src/snap_toast.dart' show SnapToast;
+export 'src/toast_pack.dart' show ToastPack;
 export 'src/toast_enums.dart'
     show ToastGravity, ToastWebPosition, ToastAnimation, ReplacementMode;
 export 'src/toast_leading.dart'